Move the `api-common` module under the `@bayou` namespace.
