+++ client/js/TopControl.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { SplitKey } from 'api-common';
+import { SplitKey } from '@bayou/api-common';
 import { TheModule as appCommon_TheModule } from 'app-common';
 import { EditorComplex } from 'doc-client';
 import { Logger } from '@bayou/see-all';
