+++ client/package.json
@@ -16,7 +16,7 @@
   "testMain": "js/tests/index.js",
 
   "dependencies": {
-    "api-common": "local",
+    "@bayou/api-common": "local",
     "app-common": "local",
     "doc-client": "local",
     "env-client": "local",
