+++ local-modules/api-client/TargetHandler.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { TargetId } from 'api-common';
+import { TargetId } from '@bayou/api-common';
 import { TFunction } from '@bayou/typecheck';
 import { Functor, MethodCacheProxyHandler } from '@bayou/util-common';
