+++ local-modules/app-common/TheModule.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { TheModule as apiCommon_TheModule } from 'api-common';
+import { TheModule as apiCommon_TheModule } from '@bayou/api-common';
 import { Codec, Registry } from '@bayou/codec';
 import { TheModule as docCommon_TheModule } from '@bayou/doc-common';
 import { TheModule as otCommon_TheModule } from '@bayou/ot-common';
